% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_range.R
\name{njoaguof_data_range_in_years}
\alias{njoaguof_data_range_in_years}
\title{Number of years spanned by the installed njoaguof data}
\usage{
njoaguof_data_range_in_years()
}
\value{
The span of `njoaguof::incident$incident_date_1`, in years.
}
\description{
The per-capita map scale is a yearly rate, so it needs to know how many
years of data it is averaging over. This derives that figure from
[njoaguof::incident] at the moment it is called.
}
\examples{
njoaguof_data_range_in_years()
}
