% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njoaguofdash-package.R
\docType{package}
\name{njoaguofdash-package}
\alias{njoaguofdash}
\alias{njoaguofdash-package}
\title{njoaguofdash: NJ OAG Use of Force data dashboard}
\description{
A dashboard application illustrating the njoaguof data package, which is a repackaging of the NJ OAG Use of Force dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tor-gu/njoaguofdash}
  \item Report bugs at \url{https://github.com/tor-gu/njoaguofdash/issues}
}

}
\author{
\strong{Maintainer}: Tor Gunston \email{tor.gunston@protonmail.com}

Authors:
\itemize{
  \item Tor Gunston \email{tor.gunston@protonmail.com}
}

}
\keyword{internal}
