% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{njoaguofdashApp}
\alias{njoaguofdashApp}
\title{NJ OAG Use Of Force Dashboard app}
\usage{
njoaguofdashApp()
}
\value{
Shiny app
}
\description{
NJ OAG Use Of Force Dashboard app
}
\examples{
\dontrun{
njoaguofdashApp()
}
}
