% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{njoaguofdashApp}
\alias{njoaguofdashApp}
\title{NJ OAG Use Of Force Dashboard app}
\usage{
njoaguofdashApp(md_dir = NULL)
}
\arguments{
\item{md_dir}{Path to render the about.md file. If omitted or NULL, will
render to a temp dir}
}
\value{
Shiny app
}
\description{
NJ OAG Use Of Force Dashboard app
}
\examples{
\dontrun{
njoaguofdashApp()
}
}
